Implemented page range selection in filters_ui.dart
- Improved book_filter.dart: minPages and maxPages use int instead of double now.
- Enhanced text when equal values are selected in RatingsDialog

+++ lib/models/book_filter.dart
@@ -19,9 +19,9 @@ class BookSearchFilter {
 
   double maxPrice;
 
-  double minPages;
+  int minPages;
 
-  double maxPages;
+  int maxPages;
 
   BookSearchFilter({
     this.authors = const [],
